pyproject.toml
[tool.poetry]
name = "bread-bot"
version = "0.3.0"
description = ""
authors = ["Kudryashov Lev LEVLLN"]
readme = "README.md"
packages = [{include = "bread_bot"}]

[tool.poetry.dependencies]
python = "^3.11"
psycopg2-binary = "^2.9.5"
sqlalchemy = "^1.4.44"
alembic = "^1.8.1"
asyncpg = "^0.27.0"
fastapi = "^0.88.0"
uvicorn = "^0.20.0"
uvloop = "^0.17.0"
httptools = "^0.5.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
requests = "^2.28.1"
opencensus = "^0.11.0"
opencensus-ext-logging = "^0.1.1"
opencensus-ext-httplib = "^0.7.5"
opencensus-ext-requests = "^0.8.0"
python-dateutil = "^2.8.2"
asynclog = "^0.1.7"
markdown = "^3.4.1"
beautifulsoup4 = "^4.11.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.5"
asgiref = "^3.5.2"
python-dotenv = "^0.21.0"
websockets = "^10.4"
httpx = "0.23.1"
idna = "^3.4"
pytest-httpx = "^0.21.2"
urllib3 = "^1.26.13"
pydantic = {version = "1.10.2", extras = ["email"]}
certifi = "2022.12.07"
pre-commit = "^3.2.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
flake8 = "^6.0.0"
aiosqlite = "^0.17.0"
coverage = "^6.5.0"
freezegun = "^1.2.2"
pytest-cov = "^4.0.0"
pytest-mock = "^3.10.0"
pytest-freezegun = "^0.4.2"
pytest-asyncio = "^0.20.2"
asynctest = "^0.13.0"
respx = "^0.20.1"
flake8-pyproject = "^1.2.2"
flake8-black = "^0.3.5"
ruff = "^0.0.259"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
target-version = "py311"
exclude = ["*migrations*"]
select = [
    "E", # pyflakes
    "F", # pycodestyle errors
    "W", # pycodestyle warnings
    "UP", # pyupgrade
    "I", # isort
    "C4", # flake8-comprehensions
    # pytest
    "PT018", # Assertion should be broken down into multiple parts
    "PT022", # No teardown in fixture {name}, use return instead of yield
]

[tool.flake8]
ignore = ["E231", "E241", "F403", "W503", "E712", "E203"]
per-file-ignores = [
    "__init__.py:F401",
]
max-line-length = 120
count = true
exclude = [".git","__pycache__","docs/source/conf.py","old","build","dist","tests"]

[tool.black]
line-length = 120
include = '\.pyi?$'
ignore = ["E203"]
preview = true
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.pytest_cache
  | \.tox
  | \.venv
  | _build
  | build
  | migrations
  | ./bread_bot/tests
)/
'''
